+++ modules/frames/focus.lua
@@ -36,13 +36,13 @@ local function StyleFocusFrame()
     -- Add Dragons etc if needed
     local frameTexture;
     if ( unitClassification == 'worldboss' or unitClassification == 'elite' ) then
-		frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIClean\\media\\UI-TargetingFrame-Elite';
+		frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIPlus\\media\\UI-TargetingFrame-Elite';
 	elseif ( unitClassification == 'rareelite' ) then
-		frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIClean\\media\\UI-TargetingFrame-Rare-Elite';
+		frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIPlus\\media\\UI-TargetingFrame-Rare-Elite';
 	elseif ( unitClassification == 'rare' ) then
-        frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIClean\\media\\UI-TargetingFrame-Rare';
+        frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIPlus\\media\\UI-TargetingFrame-Rare';
     else
-        frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIClean\\media\\UI-TargetingFrame';
+        frameTexture = 'Interface\\Addons\\ImprovedBlizzardUIPlus\\media\\UI-TargetingFrame';
 	end
 
     FocusFrame.borderTexture:SetTexture(frameTexture);