+++ modules/misc/endcap.lua
@@ -21,8 +21,8 @@ local function HideBarArt()
     @ return void
 ]]
 local function HandleEvents (self, event, ...)
-	if (event == 'ADDON_LOADED' and ... == 'ImprovedBlizzardUIClean') then
-		if (... == 'ImprovedBlizzardUIClean') then
+	if (event == 'ADDON_LOADED' and ... == 'ImprovedBlizzardUIPlus') then
+		if (... == 'ImprovedBlizzardUIPlus') then
 			if (PrimaryDB.HideActionBarArt) then
 				HideBarArt();
 			end